Moving PDLClient to builder design pattern
This PR updates the PDLClient to use the builder design pattern opening
the doors for us to easier allow for options to be passed to the PDLClient
like sandbox mode.

You can now create a PDLCLient by this:

```
let client = PDLClient::new("API_KEY").version("v5").options(<FUTURE>).build()
```

This is nice!
https://rust-unofficial.github.io/patterns/patterns/creational/builder.html